requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
sqlite3 (built-in with Python)
pandas
fpdf
streamlit-lottie
requests
